+++ book-service/src/main/java/app/book/tag/service/BOTagService.java
@@ -4,12 +4,14 @@ import app.book.api.tag.BOCreateTagRequest;
 import app.book.api.tag.BOSearchTagRequest;
 import app.book.api.tag.BOSearchTagResponse;
 import app.book.api.tag.BOUpdateTagRequest;
+import app.book.book.domain.BookTag;
 import app.book.tag.domain.Tag;
 import core.framework.db.Query;
 import core.framework.db.Repository;
 import core.framework.inject.Inject;
 import core.framework.log.Markers;
 import core.framework.util.Strings;
+import core.framework.web.exception.BadRequestException;
 import core.framework.web.exception.NotFoundException;
 
 import java.time.LocalDateTime;
@@ -21,6 +23,8 @@ import java.util.stream.Collectors;
 public class BOTagService {
     @Inject
     Repository<Tag> tagRepository;
+    @Inject
+    Repository<BookTag> bookTagRepository;
 
     public void create(BOCreateTagRequest request) {
         Tag tag = new Tag();
@@ -69,6 +73,9 @@ public class BOTagService {
     public void delete(Long id) {
         tagRepository.get(id).orElseThrow(() -> new NotFoundException(
             Strings.format("tag not found, id = {}", id), Markers.errorCode("BOOK_TAG_NOT_FOUND").getName()));
+        if (bookTagRepository.count("tag_id = ?", id) != 0) {
+            throw new BadRequestException("books have this tag, cannot delete it!", Markers.errorCode("BOOK_RELATED_TAG").getName());
+        }
         tagRepository.delete(id);
     }
 }
